refactor: [reportlog]reportlog refactor
1.add ReportLogManager, add ReportLogWorker.
2.remove Rlog, remove CommitLog.
3.move all work to thread.

Log: reportlog refactor
Task: https://pms.uniontech.com/task-view-271837.html
