+++ src/plugins/common/dfmplugin-utils/reportlog/datas/searchreportdata.cpp
