+++ src/plugins/common/dfmplugin-utils/reportlog/datas/sidebarreportdata.cpp
