+++ src/plugins/common/dfmplugin-utils/reportlog/reportlogeventreceiver.h
@@ -22,9 +22,9 @@ public:
 
     void commit(const QString &type, const QVariantMap &args);
     void handleMenuData(const QString &name, const QList<QUrl> &urlList);
-    void handleBlockMountData(const QString &id, bool result);
 
 public Q_SLOTS:
+    void handleBlockMountData(const QString &id, bool result);
     void handleMountNetworkResult(bool ret, DFMMOUNT::DeviceError err, const QString &msg);
 
 private:
