+++ src/plugins/common/dfmplugin-utils/reportlog/virtualreportlogplugin.h
@@ -15,9 +15,6 @@ namespace dfmplugin_utils {
 class VirtualReportLogPlugin : public dpf::Plugin
 {
     Q_OBJECT
-
-
-
 public:
     virtual void initialize() override;
     virtual bool start() override;
