+++ src/plugins/filemanager/dfmplugin-vault/views/createvaultview/vaultactivefinishedview.cpp
@@ -10,7 +10,7 @@
 #include "utils/fileencrypthandle.h"
 #include "utils/encryption/vaultconfig.h"
 
-#include "plugins/common/dfmplugin-utils/reportlog/rlog/datas/vaultreportdata.h"
+#include "plugins/common/dfmplugin-utils/reportlog/datas/vaultreportdata.h"
 #include <dfm-base/base/urlroute.h>
 #include <dfm-base/base/application/settings.h>
 #include <dfm-base/utils/dialogmanager.h>
